Move menu button to the Toolbar, center text
