+++ navlib/src/main/java/pl/szczodrzynski/navlib/NavToolbar.kt
@@ -1,9 +1,15 @@
 package pl.szczodrzynski.navlib
 
 import android.content.Context
+import android.graphics.drawable.LayerDrawable
 import android.util.AttributeSet
 import android.widget.ImageView
+import androidx.core.content.ContextCompat
 import com.google.android.material.appbar.MaterialToolbar
+import com.mikepenz.iconics.IconicsDrawable
+import com.mikepenz.iconics.typeface.library.navlibfont.NavLibFont
+import com.mikepenz.iconics.utils.sizeDp
+import pl.szczodrzynski.navlib.drawer.NavDrawer
 
 class NavToolbar : MaterialToolbar {
 
@@ -39,6 +45,20 @@ class NavToolbar : MaterialToolbar {
     }
 
     private fun create(attrs: AttributeSet?, defStyle: Int) {
+        val icon = ContextCompat.getDrawable(context, R.drawable.ic_menu_badge) as LayerDrawable?
+        icon?.apply {
+            mutate()
+            setDrawableByLayerId(R.id.ic_menu, IconicsDrawable(context).apply {
+                this.icon = NavLibFont.Icon.nav_menu
+                sizeDp = 24
+                colorAttr(context, R.attr.colorOnSurface)
+            })
+            setDrawableByLayerId(R.id.ic_badge, BadgeDrawable(context))
+        }
+        navigationIcon = icon
+        setNavigationOnClickListener {
+            menuClickListener?.invoke()
+        }
 
     }
 
@@ -46,6 +66,7 @@ class NavToolbar : MaterialToolbar {
     var subtitleFormatWithUnread: Int? = null
 
     var profileImageClickListener: (() -> Unit)? = null
+    var menuClickListener: (() -> Unit)? = null
 
     var profileImage
         get() = toolbarImage?.drawable
